修复窗口大小改变IP控件位置显示错误的问题， 修复子控件TreeView无法相应鼠标点击事件的问题 添加64位编译选项 添加LINK_DUILIB和DUILIB_DLL宏，在引用UILIB.h前定义，不用在使用Duilib的工程中链接Duilib的库了

+++ Core/UIRender.cpp
@@ -2119,7 +2119,7 @@ SIZE CRenderEngine::GetTextSize( HDC hDC, CPaintManagerUI* pManager , LPCTSTR ps
 	if( pstrText == NULL || pManager == NULL ) return size;
 	::SetBkMode(hDC, TRANSPARENT);
 	HFONT hOldFont = (HFONT)::SelectObject(hDC, pManager->GetFont(lpszFontID));
-	GetTextExtentPoint32(hDC, pstrText, _tcslen(pstrText) , &size);
+	GetTextExtentPoint32(hDC, pstrText, (int)_tcslen(pstrText) , &size);
 	::SelectObject(hDC, hOldFont);
 	return size;
 }
